Updates to tests.
